agent/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
Jinja2
aiofiles